+++ src/ResourceManager/Resources/Commands.Resources.Test/ScenarioTests/ResourceTests.ps1
@@ -15,6 +15,8 @@
 <#
 .SYNOPSIS
 Tests creating new resource group and a simple resource.
+.DESCRIPTION
+Smoke[Broken]Test
 #>
 function Test-CreatesNewSimpleResource
 {
@@ -46,6 +48,8 @@ function Test-CreatesNewSimpleResource
 <#
 .SYNOPSIS
 Tests creating new resource group and parent and child resources.
+.DESCRIPTION
+Smoke[Broken]Test
 #>
 function Test-CreatesNewComplexResource
 {
@@ -95,6 +99,8 @@ function Test-CreatesNewComplexResource
 <#
 .SYNOPSIS
 Tests get resources via piping from resource group
+.DESCRIPTION
+Smoke[Broken]Test
 #>
 function Test-GetResourcesViaPiping
 {
@@ -196,6 +202,8 @@ function Test-GetResourceForNonExisingResource
 <#
 .SYNOPSIS
 Tests get resources via piping from resource group
+.DESCRIPTION
+Smoke[Broken]Test
 #>
 function Test-GetResourcesViaPipingFromAnotherResource
 {
@@ -224,6 +232,8 @@ function Test-GetResourcesViaPipingFromAnotherResource
 <#
 .SYNOPSIS
 Tests moving a resource.
+.DESCRIPTION
+SmokeTest
 #>
 function Test-MoveAResource
 {
@@ -235,18 +245,26 @@ function Test-MoveAResource
 	$apiversion = "2014-04-01"
 	$resourceType = "Providers.Test/statefulResources"
 
-	# Test
-	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	New-AzureRmResourceGroup -Name $rgname2 -Location $rglocation
-	$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -ApiVersion $apiversion -Force
-	Move-AzureRmResource -ResourceId $resource.ResourceId -DestinationResourceGroupName $rgname2 -Force
+	try 
+	{
+		# Test
+		New-AzureRmResourceGroup -Name $rgname -Location $rglocation
+		New-AzureRmResourceGroup -Name $rgname2 -Location $rglocation
+		$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -ApiVersion $apiversion -Force
+		Move-AzureRmResource -ResourceId $resource.ResourceId -DestinationResourceGroupName $rgname2 -Force
 
-	$movedResource = Get-AzureRmResource -ResourceGroupName $rgname2 -ResourceName $rname -ResourceType $resourceType
+		$movedResource = Get-AzureRmResource -ResourceGroupName $rgname2 -ResourceName $rname -ResourceType $resourceType
 
-	# Assert
-	Assert-AreEqual $movedResource.Name $resource.Name
-	Assert-AreEqual $movedResource.ResourceGroupName $rgname2
-	Assert-AreEqual $movedResource.ResourceType $resource.ResourceType
+		# Assert
+		Assert-AreEqual $movedResource.Name $resource.Name
+		Assert-AreEqual $movedResource.ResourceGroupName $rgname2
+		Assert-AreEqual $movedResource.ResourceType $resource.ResourceType
+	} 
+	finally 
+	{
+		Clean-ResourceGroup $rgname
+		Clean-ResourceGroup $rgname2
+	}
 }
 
 <#
@@ -269,6 +287,8 @@ function Test-MoveResourceFailed
 <#
 .SYNOPSIS
 Tests setting a resource.
+.DESCRIPTION
+SmokeTest
 #>
 function Test-SetAResource
 {
@@ -279,25 +299,33 @@ function Test-SetAResource
 	$apiversion = "2014-04-01"
 	$resourceType = "Providers.Test/statefulResources"
 
-	# Test
-	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+	try 
+	{
+		# Test
+		New-AzureRmResourceGroup -Name $rgname -Location $rglocation
+		$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
 	
-	# Verify original value
-	$oldSku = $resource.Sku.psobject
-	$oldSkuNameProperty = $oldSku.Properties
-	Assert-AreEqual $oldSkuNameProperty.Name "name" 
-	Assert-AreEqual $resource.SKu.Name "A0" 
+		# Verify original value
+		$oldSku = $resource.Sku.psobject
+		$oldSkuNameProperty = $oldSku.Properties
+		Assert-AreEqual $oldSkuNameProperty.Name "name" 
+		Assert-AreEqual $resource.SKu.Name "A0" 
 	
-	# Set resource
-	Set-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType -Properties @{"key2" = "value2"} -Force
-	Set-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType -SkuObject @{ Name = "A1" }  -Force 
+		# Set resource
+		Set-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType -Properties @{"key2" = "value2"} -Force
+		$job = Set-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType -SkuObject @{ Name = "A1" }  -Force -AsJob
+		$job | Wait-Job
 
-	$modifiedResource = Get-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType
+		$modifiedResource = Get-AzureRmResource -ResourceGroupName $rgname -ResourceName $rname -ResourceType $resourceType
 
-	# Assert
-	Assert-AreEqual $modifiedResource.Properties.key2 "value2"
-	Assert-AreEqual $modifiedResource.Sku.Name "A1" 
+		# Assert
+		Assert-AreEqual $modifiedResource.Properties.key2 "value2"
+		Assert-AreEqual $modifiedResource.Sku.Name "A1"
+	}
+	finally
+	{
+		Clean-ResourceGroup $rgname
+	}
 }
 
 <#
@@ -329,6 +357,8 @@ function Test-SetAResourceWithPatch
 <#
 .SYNOPSIS
 Tests finding a resource.
+.DESCRIPTION
+SmokeTest
 #>
 function Test-FindAResource
 {
@@ -340,29 +370,38 @@ function Test-FindAResource
 	$apiversion = "2014-04-01"
 	$resourceType = "Providers.Test/statefulResources"
 
-	# Test
-	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	$actual = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
-	$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+	try 
+	{
+		# Test
+		New-AzureRmResourceGroup -Name $rgname -Location $rglocation
+		$actual = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+		$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 	
-	$expected = Find-AzureRmResource -ResourceType $resourceType -ResourceGroupNameContains $rgName
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+		$expected = Find-AzureRmResource -ResourceType $resourceType -ResourceGroupNameContains $rgName
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 
-	New-AzureRmResource -Name $rname2 -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
-	$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
-	Assert-AreEqual 2 @($expected).Count
+		New-AzureRmResource -Name $rname2 -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+		$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
+		Assert-AreEqual 2 @($expected).Count
 
-	$expected = Find-AzureRmResource -ResourceGroupNameEquals $rgname -ResourceNameEquals $rname
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+		$expected = Find-AzureRmResource -ResourceGroupNameEquals $rgname -ResourceNameEquals $rname
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+	}
+	finally
+	{
+		Clean-ResourceGroup $rgname
+	}
 }
 
 <#
 .SYNOPSIS
 Tests finding a resource by tag.
+.DESCRIPTION
+Smoke[Broken]Test
 #>
 function Test-FindAResource-ByTag
 {
@@ -374,29 +413,38 @@ function Test-FindAResource-ByTag
 	$apiversion = "2014-04-01"
 	$resourceType = "Providers.Test/statefulResources"
 
-	# Test
-	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	$actual = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{ScenarioTestTag = "ScenarioTestVal"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
-	$expected = Find-AzureRmResource -Tag @{ScenarioTestTag = "ScenarioTestVal"}
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+	try 
+	{
+		# Test
+		New-AzureRmResourceGroup -Name $rgname -Location $rglocation
+		$actual = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{ScenarioTestTag = "ScenarioTestVal"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+		$expected = Find-AzureRmResource -Tag @{ScenarioTestTag = "ScenarioTestVal"}
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 
-	$expected = Find-AzureRmResource -Tag @{ScenarioTestTag = $null}
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+		$expected = Find-AzureRmResource -Tag @{ScenarioTestTag = $null}
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 
-	$expected = Find-AzureRmResource -TagName "ScenarioTestTag"
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+		$expected = Find-AzureRmResource -TagName "ScenarioTestTag"
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 
-	$expected = Find-AzureRmResource -TagName "ScenarioTestTag" -TagValue "ScenarioTestVal"
-	Assert-NotNull $expected
-	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+		$expected = Find-AzureRmResource -TagName "ScenarioTestTag" -TagValue "ScenarioTestVal"
+		Assert-NotNull $expected
+		Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
+	}
+	finally
+	{
+		Clean-ResourceGroup $rgname
+	}
 }
 
 <#
 .SYNOPSIS
 Tests getting a resource with properties expanded
+.DESCRIPTION
+SmokeTest
 #>
 function Test-GetResourceExpandProperties
 {
@@ -407,16 +455,23 @@ function Test-GetResourceExpandProperties
 	$apiversion = "2014-04-01"
 	$resourceType = "Providers.Test/statefulResources"
 
-	# Test
-	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
-	$resourceGet = Get-AzureRmResource -ResourceName $rname -ResourceGroupName $rgname -ExpandProperties
+	try 
+	{
+		# Test
+		New-AzureRmResourceGroup -Name $rgname -Location $rglocation
+		$resource = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+		$resourceGet = Get-AzureRmResource -ResourceName $rname -ResourceGroupName $rgname -ExpandProperties
 
-	# Assert
-	$properties = $resourceGet.Properties.psobject
-	$keyProperty = $properties.Properties
-	Assert-AreEqual $keyProperty.Name "key"
-	Assert-AreEqual $resourceGet.Properties.key "value"
+		# Assert
+		$properties = $resourceGet.Properties.psobject
+		$keyProperty = $properties.Properties
+		Assert-AreEqual $keyProperty.Name "key"
+		Assert-AreEqual $resourceGet.Properties.key "value"
+	}
+	finally
+	{
+		Clean-ResourceGroup $rgname
+	}
 }
 
 <#
@@ -488,6 +543,8 @@ function Test-ManageResourceWithZones
 <#
 .SYNOPSIS
 Tests removing a resource.
+.DESCRIPTION
+SmokeTest
 #>
 function Test-RemoveAResource
 {
@@ -500,12 +557,15 @@ function Test-RemoveAResource
 
 	# Test
 	New-AzureRmResourceGroup -Name $rgname -Location $rglocation
-	$actual = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force
+	$job = New-AzureRmResource -Name $rname -Location $rglocation -Tags @{testtag = "testval"} -ResourceGroupName $rgname -ResourceType $resourceType -PropertyObject @{"key" = "value"} -SkuObject @{ Name = "A0" } -ApiVersion $apiversion -Force -AsJob
+	$job | Wait-Job
+	$actual = $job | Receive-Job
 	$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
 	Assert-NotNull $expected
 	Assert-AreEqual $actual.ResourceId $expected[0].ResourceId
 
-	Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname | Remove-AzureRmResource -Force
+	$job = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname | Remove-AzureRmResource -Force -AsJob
+	$job | Wait-Job
 	$expected = Find-AzureRmResource -ResourceNameContains test -ResourceGroupNameContains $rgname
 	Assert-Null $expected
 }